something, hope it works

+++ src/Nethermind/Nethermind.JsonRpc/Modules/TxPool/TransactionPoolStatus.cs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2022 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
-using System.Linq;
+using Nethermind.CoreOfCore;
 using Nethermind.TxPool;
 
 namespace Nethermind.JsonRpc.Modules.TxPool
@@ -10,8 +10,8 @@ namespace Nethermind.JsonRpc.Modules.TxPool
     {
         public TxPoolStatus(TxPoolInfo info)
         {
-            Pending = info.Pending.Sum(static t => t.Value.Count);
-            Queued = info.Queued.Sum(static t => t.Value.Count);
+            Pending = (int)info.Pending.Sum(static t => t.Value.Count);
+            Queued = (int)info.Queued.Sum(static t => t.Value.Count);
         }
 
         public int Pending { get; set; }
